refactor: 클래스 이름 변경
기존 Pair 클래스를 Pairs 클래스로 이름 변경
Pairs 클래스를 PairsRepository로 이른 변경
